requirements.txt
anyio==4.9.0
argcomplete==3.6.2
attrs==25.1.0
Beaker==1.12.1
beautifulsoup4==4.13.4
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.38.46
botocore==1.38.46
Brlapi==0.8.6
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cockpit @ file:///builddir/build/BUILD/cockpit-341-build/cockpit-341/tmp/wheel/cockpit-341-py3-none-any.whl#sha256=1dd0889753d4629153a62ed95f46f8fc113ede8579034a75d0097cf7fd002295
colorama==0.4.6
crypt_r==3.13.1
cryptography==44.0.0
cupshelpers==1.0
dasbus==1.7
dbus-python==1.3.2
dbus_next==0.2.3
distro==1.9.0
dnf==4.23.0
dnspython==2.7.0
fedora-third-party==0.10
file-magic==0.4.0
fros==1.1
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
gpg==1.24.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
humanize==4.12.0
idna==3.10
ipaddress==1.0.23
iso639==0.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langtable==0.0.68
libcomps==0.1.21
libdnf==0.74.0
louis==3.33.0
lxml==5.3.2
Mako==1.2.3
MarkupSafe==3.0.2
nftables==0.1
oauthlib==3.3.1
olefile==0.47
packaging==24.2
Paste==3.10.1
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
ply==3.11
productmd==1.45
proto-plus==1.26.1
protobuf==6.31.1
ptyprocess==0.7.0
pwquality==1.4.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.25.1
pycparser==2.20
pycrypto==2.6.1
pycups==2.0.4
pyenchant==3.2.2
PyGObject==3.50.0
pyinotify==0.9.6
pykickstart==3.62
pyOpenSSL==25.0.0
pyparsing==3.2.3
pyparted==3.13.0
pypng==0.0.21
PyQt5==5.15.11
PyQt5_sip==12.16.1
PySocks==1.7.1
python-augeas==1.1.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-pam==2.0.2
python-telegram-bot==22.2
python-whois==0.9.5
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.15.4-build/kernel-6.15.4/linux-6.15.4-200.fc42.x86_64/tools/net/ynl
qrcode==8.0
quearcode==0.4.1
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.0.0
requests-ftp==0.3.1
requests-oauthlib==2.0.0
rpds-py==0.25.0
rpm==4.20.1
rsa==4.9.1
s3transfer==0.13.0
scapy==2.6.1
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
sniffio==1.3.1
sos==4.9.1
soupsieve==2.7
systemd-python==235
torbrowser-launcher==0.3.7
typing_extensions==4.12.2
uritemplate==4.2.0
urllib3==2.3.0
xkbregistry==0.3
